news_blog/requirements.txt
Django==4.1.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
python-dotenv==0.21.0